requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
edge-tts
gunicorn
python-multipart
requests
